requirements.txt
faiss-cpu>=1.7.0
langchain>=0.0.315
pypdf==3.8.1
python-box==7.0.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
llama_cpp_python==0.2.11
docx2txt==0.8
streamlit>=1.27.0
streamlit-extras>=0.3.4